PBM-887: separate diagnostic
